TST Testing mode to bypass downloading the GTDB
For testing, downloading the GTDB multiple times is slow (and uses a lot
of bandwidth). This allows one to test the rest of the system, though.

+++ SemiBin/gtdb.py
@@ -16,20 +16,29 @@ def download_GTDB_to(logger, GTDB_dir):
     logger.info(f'Downloading GTDB to {GTDB_dir}.  It will take a while..')
     os.makedirs(GTDB_dir, exist_ok=True)
 
-    download_url = 'https://zenodo.org/record/4751564/files/GTDB_v95.tar.gz?download=1'
     download_path = os.path.join(GTDB_dir, 'GTDB_v95.tar.gz')
 
-    with requests.get(download_url, stream=True) as r:
-        m = hashlib.md5()
-        with open(download_path, 'wb') as f:
-            while True:
-                data = r.raw.read(4096)
-                if not data:
-                    break
-                f.write(data)
-                m.update(data)
-    logger.info('Download finished. Checking MD5...')
-    if m.hexdigest() == '4a70301c54104e87d5615e3f2383c8b5':
+    # This allows one to bypass the actual downloading for testing (otherwise
+    # tests take a few hours!)
+    if path.exists('__SemiBin__internal_test_GTDB_v95.tar.gz'):
+        logger.warning(f'Found __SemiBin__internal_test_GTDB_v95.tar.gz')
+        logger.warning(f'Will copy it to {download_path} and use it as GTDB tarball')
+        shutil.copy2('__SemiBin__internal_test_GTDB_v95.tar.gz', download_path)
+        md5_ok = True
+    else:
+        download_url = 'https://zenodo.org/record/4751564/files/GTDB_v95.tar.gz?download=1'
+        with requests.get(download_url, stream=True) as r:
+            m = hashlib.md5()
+            with open(download_path, 'wb') as f:
+                while True:
+                    data = r.raw.read(4096)
+                    if not data:
+                        break
+                    f.write(data)
+                    m.update(data)
+        logger.info('Download finished. Checking MD5...')
+        md5_ok = m.hexdigest() == '4a70301c54104e87d5615e3f2383c8b5'
+    if md5_ok:
         try:
             tar = tarfile.open(download_path, "r:gz")
             file_names = tar.getnames()
